finished auth v1
refactored routes

now auth is needed to every request

+++ api/src/utils/route_meta.py
@@ -1,6 +1,9 @@
 from pydantic import BaseModel
 from typing import TypeVar
-from fastapi import APIRouter, HTTPException
+from fastapi import APIRouter, HTTPException, Header, Depends
+
+from api.src.auth.auth import Auth
+from api.src.auth.bearer import Bearer
 
 
 class RouteMeta(type):
@@ -22,17 +25,23 @@ class RouteMeta(type):
                 'id': 'id',
                 'name': 'm_name',
                 'cat_id': 'category',
-                'user_id': 'user'
             }
 
-            key_path = f'{{{key_dict.get(key[0])}}}'
-            for k in key[1:]:
-                key_path += f'/{{{key_dict.get(k)}}}'
+            key_path = tuple([k for k in key if k != 'user_id'])
+            key_path_without_user = f'{{{key_dict.get(key_path[0])}}}'
+            for k in key_path[1:]:
+                key_path_without_user += f'/{{{key_dict.get(k)}}}'
 
             if 'get' in methods:
-                @router.get(f'/api/{route_base_name}/{key_path}', tags=[tag])
+                @router.get(f'/api/{route_base_name}/{key_path_without_user}',
+                            dependencies=[Depends(Bearer())], tags=[tag])
                 async def get(id: int = None, m_name: str = None,
-                              category: int = None, user: int = None):
+                              category: int = None, authorization: str = Header(None)):
+                    token = authorization.split(" ")[1]
+                    payload = Auth.decode(token)
+                    if payload is None:
+                        raise HTTPException(status_code=401, detail="Invalid token")
+                    user = payload['user_id']
                     args = list(filter(lambda x: x is not None, [id, user, m_name, category]))
                     ret = dao.get(*args)
                     if ret is None:
@@ -41,17 +50,27 @@ class RouteMeta(type):
             if 'add' in methods:
                 url = f'/api/{route_base_name}'
 
-                @router.post(url, tags=[tag])
-                async def post(mod: model, id: int = None, user: int = None):
-                    args = list(filter(lambda x: x is not None, [id, user]))
-                    added = dao.add(*args, mod) if len(args) != 0 else dao.add(mod)
+                @router.post(url, dependencies=[Depends(Bearer())], tags=[tag])
+                async def post(mod: model, authorization: str = Header(None)):
+                    token = authorization.split(" ")[1]
+                    payload = Auth.decode(token)
+                    if payload is None:
+                        raise HTTPException(status_code=401, detail="Invalid token")
+                    mod.user_id = payload['user_id']
+                    added = dao.add(mod)
                     if added is None:
                         raise HTTPException(status_code=404, detail="Item not found or already exists")
                     return added
             if 'update' in methods:
-                @router.put(f'/api/{route_base_name}/{key_path}', tags=[tag])
+                @router.put(f'/api/{route_base_name}/{key_path_without_user}', dependencies=[Depends(Bearer())],
+                            tags=[tag])
                 async def put(mod: model, id: int = None, m_name: str = None,
-                              category: int = None, user: int = None):
+                              category: int = None, authorization: str = Header(None)):
+                    token = authorization.split(" ")[1]
+                    payload = Auth.decode(token)
+                    if payload is None:
+                        raise HTTPException(status_code=401, detail="Invalid token")
+                    user = payload['user_id']
                     args = list(filter(lambda x: x is not None, [id, user, m_name, category]))
                     if dao.get(*args) is None:
                         raise HTTPException(status_code=404, detail="Item not found")
@@ -60,20 +79,28 @@ class RouteMeta(type):
                         raise HTTPException(status_code=404, detail="Internal Server Error")
                     return updated
             if 'remove' in methods:
-                @router.delete(f'/api/{route_base_name}/{key_path}', tags=[tag])
+                @router.delete(f'/api/{route_base_name}/{key_path_without_user}',
+                               dependencies=[Depends(Bearer())], tags=[tag])
                 async def delete(id: int = None, m_name: str = None,
-                                 category: int = None, user: int = None):
+                                 category: int = None, authorization: str = Header(None)):
+                    token = authorization.split(" ")[1]
+                    payload = Auth.decode(token)
+                    if payload is None:
+                        raise HTTPException(status_code=401, detail="Invalid token")
+                    user = payload['user_id']
                     args = list(filter(lambda x: x is not None, [id, user, m_name, category]))
                     if dao.get(*args) is None:
                         raise HTTPException(status_code=404, detail="Item not found")
                     return dao.remove(*args)
             if 'get_all' in methods:
 
-                @router.get(f'/api/all/{route_base_name}/{{user}}', tags=[tag])
-                async def get_all(id: int = None, m_name: str = None,
-                                  category: int = None, user: int = None):
-                    args = list(filter(lambda x: x is not None, [id, user, m_name, category]))
-                    ret = dao.get_all(*args)
+                @router.get(f'/api/all/{route_base_name}', dependencies=[Depends(Bearer())], tags=[tag])
+                async def get_all(authorization: str = Header(None)):
+                    token = authorization.split(" ")[1]
+                    payload = Auth.decode(token)
+                    if payload is None:
+                        raise HTTPException(status_code=401, detail="Invalid token")
+                    ret = dao.get_all(payload['user_id'])
                     if ret is None:
                         raise HTTPException(status_code=404, detail="Item not found")
                     return ret
